Oops, missed a couple.
Reviewed-on: https://code-review.googlesource.com/c/re2/+/43032
(cherry picked from commit f1a1156294f0337beb75518da3606091f3905fb1)

+++ re2/testing/dfa_test.cc
@@ -299,7 +299,7 @@ ReverseTest reverse_tests[] = {
 
 TEST(DFA, ReverseMatch) {
   int nfail = 0;
-  for (int i = 0; i < arraysize(reverse_tests); i++) {
+  for (size_t i = 0; i < arraysize(reverse_tests); i++) {
     const ReverseTest& t = reverse_tests[i];
     Regexp* re = Regexp::Parse(t.regexp, Regexp::LikePerl, NULL);
     ASSERT_TRUE(re != NULL);
@@ -355,7 +355,7 @@ CallbackTest callback_tests[] = {
 
 TEST(DFA, Callback) {
   int nfail = 0;
-  for (int i = 0; i < arraysize(callback_tests); i++) {
+  for (size_t i = 0; i < arraysize(callback_tests); i++) {
     const CallbackTest& t = callback_tests[i];
     Regexp* re = Regexp::Parse(t.regexp, Regexp::LikePerl, NULL);
     ASSERT_TRUE(re != NULL);
